local/requirements.txt
websocket-client==1.2.1
upnpclient==1.0.3
requests==2.26.0
